+++ pyrolite/util/synthetic.py
@@ -4,6 +4,64 @@ Utility functions for creating synthetic (geochemical) data.
 import numpy as np
 import pandas as pd
 from ..comp.codata import ilr, inverse_ilr
+from ...geochem.norm import get_reference_composition
+
+
+def example_spider_data(
+    start="EMORB_SM89",
+    norm_to="PM_PON",
+    nobs=120,
+    noise_level=0.5,
+    offsets={"Eu": 0.5},
+    units="ppm",
+):
+    """
+    Generate some random data for demonstrating spider plots.
+
+    By default, this generates a composition based around EMORB, normalised
+    to Primitive Mantle.
+
+    Parameters
+    -----------
+    start : :class:`str`
+        Composition to start with.
+    norm_to : :class:`str`
+        Composition to normalise to. Can optionally specify :code:`None`.
+    nobs : :class:`int`
+        Number of observations to include (index length).
+    noise_level : :class:`float`
+        Log-units of noise (1sigma).
+    offsets : :class:`dict`
+        Dictionary of offsets in log-units (e.g. here by default offsets Eu by 0.8 log
+        units).
+    units : :class:`str`
+        Units to use before conversion. Should have no effect other than reducing
+        calculation times if `norm_to` is :code:`None`.
+
+    Returns
+    --------
+    df : :class:`pandas.DataFrame`
+        Dataframe of example synthetic data.
+    """
+
+    ref = get_reference_composition(start)
+    ref.set_units(units)
+    df = ref.comp.pyrochem.compositional
+    if norm_to is not None:
+        df = df.pyrochem.normalize_to(norm_to, units=units)
+    start = df.applymap(np.log)
+    nindex = df.columns.size
+
+    x = np.arange(nindex)
+    y = np.tile(start.values, nobs).reshape(nobs, nindex)
+    y += np.random.normal(0, noise_level / 2.0, size=(nobs, nindex))  # noise
+    y += np.random.normal(0, noise_level, size=(1, nobs)).T  # random pattern offset
+
+    syn_df = pd.DataFrame(y, columns=df.columns)
+    for element, offset in offsets.items():
+        syn_df[element] += offset  # significant offset for Eu anomaly
+    syn_df = syn_df.applymap(np.exp)
+    return syn_df
 
 
 def random_cov_matrix(dim, sigmas=None, validate=False, seed=None):
@@ -139,7 +197,9 @@ def random_composition(
     # covariance
     if cov is None:
         if D != 1:
-            cov = random_cov_matrix(D - 1, sigmas=np.abs(mean) * 0.1, seed=seed)  # 10% sigmas
+            cov = random_cov_matrix(
+                D - 1, sigmas=np.abs(mean) * 0.1, seed=seed
+            )  # 10% sigmas
         else:
             cov = np.array([[1]])
 
