+++ docs/source/usage/examples/_auto/plotting/spider.py
@@ -6,36 +6,34 @@ import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 
-# sphinx_gallery_thumbnail_number = 5
+# sphinx_gallery_thumbnail_number = 4
 
 ########################################################################################
-# Here we'll set up an example which uses EMORB as a starting point:
-#
-from pyrolite.geochem.norm import get_reference_composition
-
-ref = get_reference_composition("EMORB_SM89")  # EMORB composition as a starting point
-ref.set_units("ppm")
-df = ref.comp.pyrochem.compositional
-########################################################################################
-# Basic spider plots are straightforward to produce:
-import pyrolite.plot
-
-df.pyroplot.spider(color="k")
-plt.show()
-########################################################################################
-# Typically we'll normalise trace element compositions to a reference composition
+# Here we'll set up an example which uses EMORB as a starting point. Typically we'll
+# normalise trace element compositions to a reference composition
 # to be able to link the diagram to 'relative enrichement' occuring during geological
-# processes:
+# processes, so here we're normalising to a Primitive Mantle composition first.
+# We're here taking this normalised composition and adding some noise in log-space to
+# generate multiple compositions about this mean (i.e. a compositional distribution).
+# For simplicility, this is handlded by
+# :func:`~pyrolite.util.synthetic.example_spider_data`:
 #
-normdf = df.pyrochem.normalize_to("PM_PON", units="ppm")
-ax = normdf.pyroplot.spider(color="k", unity_line=True)
-ax.set_ylabel('X / $X_{Primitive Mantle}$')
-plt.show()
+from pyrolite.util.synthetic import example_spider_data
+
+normdf = example_spider_data(start="EMORB_SM89", norm_to="PM_PON")
 
 ########################################################################################
 # .. seealso:: `Normalisation <../geochem/normalization.html>`__
 #
 
+########################################################################################
+# Basic spider plots are straightforward to produce:
+#
+import pyrolite.plot
+
+ax = normdf.pyroplot.spider(color="0.5", alpha=0.5, unity_line=True, figsize=(10, 4))
+ax.set_ylabel("X / $X_{Primitive Mantle}$")
+plt.show()
 ########################################################################################
 # The default ordering here follows that of the dataframe columns, but we typically
 # want to reorder these based on some physical ordering. A :code:`index_order` keyword
@@ -44,46 +42,37 @@ plt.show()
 # :func:`pyrolite.geochem.ind.order_incompatibility`:
 from pyrolite.geochem.ind import by_incompatibility
 
-ax = normdf.pyroplot.spider(color="k", unity_line=True, index_order=by_incompatibility)
-ax.set_ylabel('X / $X_{Primitive Mantle}$')
+ax = normdf.pyroplot.spider(
+    color="k",
+    alpha=0.1,
+    unity_line=True,
+    index_order=by_incompatibility,
+    figsize=(10, 4),
+)
+ax.set_ylabel("X / $X_{Primitive Mantle}$")
 plt.show()
 ########################################################################################
 # The spiderplot can be extended to provide visualisations of ranges and density via the
-# various modes. First let's take this composition and add some noise in log-space to
-# generate multiple compositions about this mean (i.e. a compositional distribution):
+# various modes. We could now plot the range of compositions as a filled range:
 #
[…]
