### 2020-03-04 2833e24
Reduce function reference for set_units in norm example
