### 2020-03-04 2ca2ad7
Update util.synthetic tests
