### 2020-03-04 6d138a3
Update REE_v_radii docs example
