### 2020-03-05 460ede8
Update REE to default to dropPm=True
