### 2020-03-05 ba72d6c
Update logy handling for spider and REE_v_radii
