### 2020-03-05 b8c3e77
Add filter to plot.REE
