### 2020-03-05 6ec02cc
Add minimum figure dimension to util.plot.axes
