+++ docs/source/usage/examples/_auto/geochem/normalization.py
@@ -28,9 +28,8 @@ fig, ax = plt.subplots(1)
 for name, ref in list(all_reference_compositions().items())[::2]:
     if name != "Chondrite_PON":
         ref.set_units("ppm")
-        ref.comp.pyrochem.REE.pyrochem.normalize_to(CI, units="ppm").pyroplot.REE(
-            unity_line=True, ax=ax, label=name
-        )
+        df = ref.comp.pyrochem.REE.pyrochem.normalize_to(CI, units="ppm")
+        df.pyroplot.REE(unity_line=True, ax=ax, label=name)
 
 ax.set_ylabel("X/X$_{Chondrite}$")
 ax.legend(frameon=False, facecolor=None, loc="upper left", bbox_to_anchor=(1.0, 1.0))
