### 2020-03-18 1ee7bf4
Update test for OP_constants
