### 2020-03-18 2a19a30
Minor updates for util.math docstrings
