### 2020-03-19 1cb1a8a
Add Kaarel to contributors.rst
