+++ docs/source/dev/contributors.rst
@@ -10,3 +10,4 @@ comments, testing, bug reports, feature requests.
 * Adam Bath
 * Yajing Mao
 * `Justin Gosses <https://github.com/JustinGOSSES>`__
+* `Kaarel Mand <https://github.com/kaarelmand>`__
