### 2020-04-06 ad9ccc7
Split util.math to create util.lambdas
