+++ pyrolite/util/lambdas.py
@@ -1,0 +1,269 @@
+import numpy as np
+from sympy.solvers.solvers import nsolve
+from sympy import symbols, var
+from functools import partial
+import scipy
+from .meta import update_docstring_references
+import logging
+
+logging.getLogger(__name__).addHandler(logging.NullHandler())
+logger = logging.getLogger(__name__)
+
+@update_docstring_references
+def OP_constants(xs, degree=3, rounding=None, tol=10 ** -14):
+    r"""
+    Finds the parameters
+    :math:`(\beta_0), (\gamma_0, \gamma_1), (\delta_0, \delta_1, \delta_2)` etc.
+    for constructing orthogonal polynomial functions `f(x)` over a fixed set of values
+    of independent variable `x`.
+    Used for obtaining lambda values for dimensional reduction of REE data [#ref_1]_.
+
+    Parameters
+    ----------
+    xs : :class:`numpy.ndarray`
+        Indexes over which to generate the orthogonal polynomials.
+    degree : :class:`int`
+        Maximum polynomial degree. E.g. 2 will generate constant, linear, and quadratic
+        polynomial components.
+    tol : :class:`float`
+        Convergence tolerance for solver.
+    rounding : :class:`int`
+        Precision for the orthogonal polynomial coefficents.
+
+    Returns
+    ---------
+    :class:`list`
+        List of tuples corresponding to coefficients for each of the polynomial
+        components. I.e the first tuple will be empty, the second will contain a single
+        coefficient etc.
+
+    Notes
+    -----
+        Parameters are used to construct orthogonal polymomials of the general form:
+
+        .. math::
+
+            f(x) &= a_0 \\
+            &+ a_1 * (x - \beta) \\
+            &+ a_2 * (x - \gamma_0) * (x - \gamma_1) \\
+            &+ a_3 * (x - \delta_0) * (x - \delta_1) * (x - \delta_2) \\
+
+    See Also
+    ---------
+    :func:`~pyrolite.util.math.lambdas`
+    :func:`~pyrolite.geochem.transform.lambda_lnREE`
+
+    References
+    -----------
+    .. [#ref_1] O’Neill HSC (2016) The Smoothness and Shapes of Chondrite-normalized
+           Rare Earth Element Patterns in Basalts. J Petrology 57:1463–1508.
+           doi: `10.1093/petrology/egw047 <https://dx.doi.org/10.1093/petrology/egw047>`__
+    """
+    xs = np.array(xs)
+    x = var("x")
+    params = []
+    for d in range(degree):
+        ps = symbols("{}0:{}".format(chr(945 + d), d))
+        logger.debug("Generating {} DIM {} equations for {}.".format(d, d, ps))
+        if d:
+            eqs = []
+            for _deg in range(d):
+                q = 1
+                if _deg:
+                    q = x ** _deg
+                for p in ps:
+                    q *= x - p
+                eqs.append(q)
+            sums = []
+            for q in eqs:
+                sumq = 0.0
+                for xi in xs:
+                    sumq += q.subs(dict(x=xi))
+                sums.append(sumq)
+
+            guess = np.linspace(np.nanmin(xs), np.nanmax(xs), d + 2)[1:-1]
+            result = nsolve(sums, ps, list(guess), tol=tol)
+            if rounding is not None:
+                result = np.around(np.array(result, dtype=np.float), decimals=rounding)
+            params.append(tuple(result))
+        else:
+            params.append(())  # first parameter
+    return params
+
+
+def lambda_poly(x, ps):
+    """
+    Polynomial lambda_n(x) given parameters ps with len(ps) = n.
+
+    Parameters
+    -----------
+    x : :class:`numpy.ndarray`
[…]
