### 2020-04-06 9ef1dc6
Update tests, references to lambdas
