+++ pyrolite/geochem/__init__.py
@@ -487,8 +487,8 @@ class pyrochem(object):
         See Also
         ---------
         :func:`~pyrolite.geochem.ind.get_ionic_radii`
-        :func:`~pyrolite.util.math.lambdas`
-        :func:`~pyrolite.util.math.OP_constants`
+        :func:`~pyrolite.util.lambdas.lambdas`
+        :func:`~pyrolite.util.lambdas.OP_constants`
         :func:`~pyrolite.plot.REE_radii_plot`
         """
         return transform.lambda_lnREE(
