### 2020-04-06 5fddc3d
Clear Docs Examples Dir
