+++ docs/source/examples/geochem/mineral_lattice.py
@@ -1,152 +1,0 @@
-"""
-Lattice Strain Calculations
-------------------------------
-
-.. note::
-
-    This example follows that given during a Institute of Advanced Studies Masterclass
-    with Jon Blundy at the University of Western Australia on the 29\ :sup:`th` April
-    2019, and is used here with permission.
-
-
-Pyrolite includes a function for calculating relative lattice strain [#ref_1]_, which
-together with the tables of Shannon ionic radii and Young's modulus approximations for
-silicate and oxide cationic sites enable relatively simple calculation of ionic
-partitioning in common rock forming minerals.
-
-This example below uses previously characterised calcium and sodium partition
-coefficients between plagioclase (:math:`CaAl_2Si_2O_8 - NaAlSi_3O8`) and silicate melt
-to estimate partitioning for other cations based on their ionic radii.
-
-A model parameterised using sodium and calcium partition coefficients [#ref_2]_ is then
-used to estimate the partitioning for lanthanum into the trivalent site (largely
-occupied by :math:`Al^{3+}`), and extended to other trivalent cations (here, the Rare
-Earth Elements). The final section of the example highlights the mechanism which
-generates plagioclase's hallmark 'europium anomaly', and the effects of variable
-europium oxidation state on bulk europium partitioning.
-"""
-import numpy as np
-import matplotlib.pyplot as plt
-from pyrolite.geochem.ind import REE, get_ionic_radii
-from pyrolite.mineral.lattice import strain_coefficient
-# sphinx_gallery_thumbnail_number = 3
-########################################################################################
-# First, we need to define some of the necessary parameters including temperature, the Young's
-# moduli for the :math:`X^{2+}` and :math:`X^{3+}` sites in plagioclase (:math:`E_2`,
-# :math:`E_3`), and some reference partition coefficients and radii for calcium and
-# sodium:
-D_Na = 1.35  # Partition coefficient Plag-Melt
-D_Ca = 4.1  # Partition coefficient Plag-Melt
-Tc = 900  # Temperature, °C
-Tk = Tc + 273.15  # Temperature, K
-E_2 = 120 * 10 ** 9  # Youngs modulus for 2+ site, Pa
-E_3 = 135 * 10 ** 9  # Youngs modulus for 3+ site, Pa
-r02, r03 = 1.196, 1.294  # fictive ideal cation radii for these sites
-rCa = get_ionic_radii("Ca", charge=2, coordination=8)
-rLa = get_ionic_radii("La", charge=3, coordination=8)
-########################################################################################
-# We can calculate and plot the partitioning of :math:`X^{2+}` cations relative to
-# :math:`Ca^{2+}` at a given temperature using their radii and the lattice strain function:
-#
-fontsize = 8
-fig, ax = plt.subplots(1)
-
-site2labels = ["Na", "Ca", "Eu", "Sr"]
-# get the Shannon ionic radii for the elements in the 2+ site
-site2radii = [
-    get_ionic_radii("Na", charge=1, coordination=8),
-    *[get_ionic_radii(el, charge=2, coordination=8) for el in ["Ca", "Eu", "Sr"]],
-]
-# plot the relative paritioning curve for cations in the 2+ site
-site2Ds = D_Ca * np.array(
-    [strain_coefficient(rCa, rx, r0=r02, E=E_2, T=Tk) for rx in site2radii]
-)
-ax.scatter(site2radii, site2Ds, color="g", label="$X^{2+}$ Cations")
[…]
