### 2020-04-06 5a81305
Pin sphinx_gallery verison
