### 2020-04-07 4f656cf
Update logo imports
