### 2020-04-07 68afeac
Update examples. tutorial material
