+++ docs/source/examples/comp/compositional_data.py
@@ -1,0 +1,231 @@
+"""
+Compositional Data?
+--------------------
+"""
+
+########################################################################################
+# pyrolite comes with a few datasets from Aitchison (1984) built in which we can use
+# as examples:
+#
+import numpy as np
+import pandas as pd
+import matplotlib.pyplot as plt
+from pyrolite.plot import pyroplot
+from pyrolite.data.Aitchison import load_kongite
+
+df = load_kongite()
+########################################################################################
+# For compositional data, everything is relative (thanks to the closure property), so we
+# tend to use ratios to express differences or changes between things. However,
+# if we make incorrect assumptions about the nature of our data, we can get some incorrect
+# answers. Say you want to know the average ratio between A and B:
+#
+A_on_B = df["A"] / df["B"]
+A_on_B.mean()  # 2.8265837788402983
+########################################################################################
+# Equally, you could have chosen to calculate the average ratio between B and A
+#
+B_on_A = df["B"] / df["A"]
+B_on_A.mean()  # 0.4709565704852008
+########################################################################################
+# You expect these to be invertable, such that A_on_B = 1 / B_on_A; but not so!
+#
+A_on_B.mean() / (1 / B_on_A.mean())  # 1.3311982026717262
+########################################################################################
+# Similarly, the relative variances are different:
+#
+np.std(A_on_B) / A_on_B.mean()  # 0.6295146309597085
+np.std(B_on_A) / B_on_A.mean()  # 0.5020948201979953
+########################################################################################
+# This improves when using logratios in place of simple ratios, prior to exponentiating means
+#
+logA_on_B = (df["A"] / df["B"]).apply(np.log)
+logB_on_A = (df["B"] / df["A"]).apply(np.log)
+########################################################################################
+# The logratios are invertible:
+#
+np.exp(logA_on_B.mean())  # 2.4213410747400514
+1 / np.exp(logB_on_A.mean())  # 2.421341074740052
+########################################################################################
+# The logratios also have the same variance:
+(np.std(logA_on_B) / logA_on_B.mean()) ** 2  # 0.36598579018127086
+(np.std(logB_on_A) / logB_on_A.mean()) ** 2  # 0.36598579018127086
+
+########################################################################################
+# These peculiarities result from incorrect assumptions regarding the distribution of the
+# data: ratios of compositional components are typically *lognormally* distributed, rather
+# than *normally* distributed, and the compositional components themselves commonly have
+# a `Poisson distribution <https://en.wikipedia.org/wiki/Poisson_distribution>`__ .
+# These distributions contrast significantly with the normal distribution at the core of
+# most statistical tests. We can compare distributions with similar means and variances
+# but different forms, and note that the normal distribution has one immediate
+# failure, in that it has non-zero probability density below 0, and we know that you can't
+# have negative atoms!
+#
+from scipy.stats import norm, poisson, lognorm
+
+means = [[10, 10], [10, 20], [20, 100], [1000, 50]]
[…]
