+++ docs/source/tutorials/ternary_density.py
@@ -49,7 +49,7 @@ coords, H, data = ternary_heatmap(
 import matplotlib.pyplot as plt
 import pyrolite.plot
 from pyrolite.util.math import flattengrid
-from pyrolite.util.plot import axes_to_ternary, share_axes
+from pyrolite.util.plot.axes import axes_to_ternary, share_axes
 
 fig, ax = plt.subplots(3, 2, figsize=(6, 9))
 ax = ax.flat

[…]
