### 2020-04-07 f529b51
Update dependency to unpin sphinx_gallery
