### 2020-04-07 1930b40
Update docs patch for sphinx_gallery
