+++ docs/source/docpatch.py
@@ -1,312 +1,0 @@
-import re
-import os
-import sys
-import copy
-import ast
-from functools import partial
-from io import StringIO
-import logging
-from pathlib import Path
-from sphinx_gallery import gen_rst
-from sphinx_gallery import binder
-from sphinx_gallery.gen_rst import *
-from sphinx_gallery.binder import *
-from sphinx_gallery import scrapers
-import matplotlib
-import matplotlib.pyplot as plt
-from matplotlib import _pylab_helpers
-from pyrolite.util.plot import save_figure
-
-_si = """
-.. image:: /%s
-    :width: 80 %
-    :align: center
-"""
-html_header = """.. only:: builder_html
-
-.. raw:: html
-
-    {0}\n        <br />\n        <br />"""
-
-
-def alt_matplotlib_scraper(block, block_vars, gallery_conf, **kwargs):
-    """Patched matplotlib scraper which won't close figures.
-
-    Parameters
-    ----------
-    block : tuple
-        A tuple containing the (label, content, line_number) of the block.
-    block_vars : dict
-        Dict of block variables.
-    gallery_conf : dict
-        Contains the configuration of Sphinx-Gallery
-    **kwargs : dict
-        Additional keyword arguments to pass to
-        :meth:`~matplotlib.figure.Figure.savefig`, e.g. ``format='svg'``.
-        The ``format`` kwarg in particular is used to set the file extension
-        of the output file (currently only 'png' and 'svg' are supported).
-
-    Returns
-    -------
-    rst : str
-        The ReSTructuredText that will be rendered to HTML containing
-        the images. This is often produced by :func:`~sphinx_galleryscrapers.figure_rst`.
-    """
-    lbl, cnt, ln = block
-    image_path_iterator = block_vars["image_path_iterator"]
-
-    image_paths = []
-    figs = [m.canvas.figure for m in _pylab_helpers.Gcf.get_all_fig_managers()]
-    fltr = "[^a-zA-Z0-9]+fig,\s??|[^a-zA-Z0-9]+fig\s??|[^a-zA-Z0-9]+figure\s?|plt.show()|plt.gcf()"
-    if figs and re.search(fltr, cnt):  # where figure or plt.show is called
-        for fig, image_path in zip([figs[-1]], image_path_iterator):
-            to_rgba = matplotlib.colors.colorConverter.to_rgba
-            save_figure(
-                fig,
-                save_at=Path(image_path).parent,
-                name=Path(image_path).stem,
-                save_fmts=["png"],
-                **kwargs.copy()
-            )
-            image_paths.append(image_path)
-
-    return scrapers.figure_rst(image_paths, gallery_conf["src_dir"])
-
-
-def alt_gen_binder_rst(
-    fpath, binder_conf, gallery_conf, img="https://mybinder.org/badge_logo.svg"
-):
-    """Generate the RST + link for the Binder badge.
-    Parameters
-    ----------
-    fpath: str
-        The path to the `.py` file for which a Binder badge will be generated.
-    binder_conf: dict or None
-        If a dictionary it must have the following keys:
-        'binderhub_url'
-            The URL of the BinderHub instance that's running a Binder service.
-        'org'
-            The GitHub organization to which the documentation will be pushed.
-        'repo'
-            The GitHub repository to which the documentation will be pushed.
-        'branch'
-            The Git branch on which the documentation exists (e.g., gh-pages).
-        'dependencies'
-            A list of paths to dependency files that match the Binderspec.
-    Returns
-    -------
-    rst : str
-        The reStructuredText for the Binder badge that links to this file.
-    """
-    binder_conf = check_binder_conf(binder_conf)
[…]
