### 2020-04-07 f788c21
Remove default export dpi
