+++ docs/source/examples/plotting/parallel.rst
@@ -128,7 +128,7 @@ of spurious correlation:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  5.955 seconds)
+   **Total running time of the script:** ( 0 minutes  1.124 seconds)
 
 
 .. _sphx_glr_download_examples_plotting_parallel.py:
