+++ docs/source/examples/comp/sg_execution_times.rst
@@ -5,10 +5,10 @@
 
 Computation times
 =================
-**00:24.946** total execution time for **examples_comp** files:
+**00:13.586** total execution time for **examples_comp** files:
 
 +---------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_examples_comp_compositional_data.py` (``compositional_data.py``) | 00:17.274 | 0.0 MB |
+| :ref:`sphx_glr_examples_comp_compositional_data.py` (``compositional_data.py``) | 00:08.415 | 0.0 MB |
 +---------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_examples_comp_logratios.py` (``logratios.py``)                   | 00:07.672 | 0.0 MB |
+| :ref:`sphx_glr_examples_comp_logratios.py` (``logratios.py``)                   | 00:05.171 | 0.0 MB |
 +---------------------------------------------------------------------------------+-----------+--------+
