+++ docs/source/tutorials/sg_execution_times.rst
@@ -5,10 +5,10 @@
 
 Computation times
 =================
-**00:34.917** total execution time for **tutorials** files:
+**00:10.824** total execution time for **tutorials** files:
 
 +-----------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_tutorials_logo.py` (``logo.py``)                       | 00:26.927 | 0.0 MB |
+| :ref:`sphx_glr_tutorials_logo.py` (``logo.py``)                       | 00:07.227 | 0.0 MB |
 +-----------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_tutorials_ternary_density.py` (``ternary_density.py``) | 00:07.991 | 0.0 MB |
+| :ref:`sphx_glr_tutorials_ternary_density.py` (``ternary_density.py``) | 00:03.597 | 0.0 MB |
 +-----------------------------------------------------------------------+-----------+--------+
