+++ docs/source/tutorials/logo.rst
@@ -244,7 +244,7 @@ We can also do this for individual samples, and estimate percentile contours:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  26.927 seconds)
+   **Total running time of the script:** ( 0 minutes  7.227 seconds)
 
 
 .. _sphx_glr_download_tutorials_logo.py:
