+++ docs/source/examples/util/TAS.rst
@@ -200,7 +200,7 @@ TAS diagram is illustrated:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  2.792 seconds)
+   **Total running time of the script:** ( 0 minutes  0.585 seconds)
 
 
 .. _sphx_glr_download_examples_util_TAS.py:
