+++ docs/source/examples/plotting/heatscatter.rst
@@ -107,7 +107,7 @@ visualisations for the same data:
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 0 minutes  8.461 seconds)
+   **Total running time of the script:** ( 0 minutes  3.503 seconds)
 
 
 .. _sphx_glr_download_examples_plotting_heatscatter.py:
