+++ docs/source/examples/geochem/mineral_mindb.rst
@@ -48,7 +48,7 @@ functions:
  .. code-block:: none
 
 
-    ['amphibole', 'spinel', 'pyroxene', 'garnet', 'mica', 'epidote', 'olivine', 'feldspar']
+    ['amphibole', 'pyroxene', 'mica', 'olivine', 'spinel', 'garnet', 'epidote', 'feldspar']
 
 
 
@@ -67,7 +67,7 @@ functions:
  .. code-block:: none
 
 
-    ['morimotoite', 'ferroaluminoceladonite', 'polylithionite', 'magnesioreibeckite', 'grossular', 'almandine', 'clintonite', 'chromphyllite', 'anorthite', 'magnesioarfvedsonite', 'ferrohornblende', 'chromoceladonite', 'arvedsonite', 'jadeite', 'annite', 'magnesiohastingsite', 'edenite', 'uvarovite', 'siderophyllite', 'ferropargasite', 'richterite', 'forsterite', 'tremolite', 'namansilite', 'andradite', 'eastonite', 'esseneite', 'phengite', 'piemontite', 'eckermanite', 'glaucophane', 'microcline', 'hedenbergite', 'margarite', 'fayalite', 'liebenbergite', 'tephroite', 'spessartine', 'tschermakite', 'ferroceladonite', 'ferrokatophorite', 'ferroeckermanite', 'ferrotschermakite', 'celadonite', 'ferroedenite', 'katophorite', 'riebeckite', 'anthopyllite', 'muscovite', 'phlogopite', 'ferrorichterite', 'enstatite', 'aluminoceladonite', 'ferrosilite', 'magnesiochromite', 'chromite', 'hercynite', 'spinel', 'pyrope', 'magnesioferrite', 'aegirine', 'magnetite', 'hastingsite', 'kosmochlor', 'albite', 'diopside', 'manganiceladonite', 'clinozoisite', 'paragonite', 'winchite', 'ferrokaersutite', 'barroisite', 'allanite', 'majorite', 'epidote', 'gedrite', 'kaersutite', 'pargasite', 'johannsenite', 'spodumene', 'trilithionite', 'taramite', 'magnesiohornblende']
[…]

[…]
