### 2020-04-07 6dae81b
Add local docs_resolv to patch
