+++ docs/source/patch/_sphinx_gallery_patch.py
@@ -1,4 +1,4 @@
 from .sphinx_gallery_gen_rst import *
 from .sphinx_gallery_binder import *
 from .sphinx_gallery_scrapers import *
-#from .sphinx_gallery_docs_resolv import *
+from .sphinx_gallery_docs_resolv import *
