### 2020-04-07 2c4b466
Update Sphinx Gallery Patch Import Order
