+++ docs/source/patch/_sphinx_gallery_patch.py
@@ -1,4 +1,4 @@
-from .sphinx_gallery_gen_rst import *
-from .sphinx_gallery_binder import *
-from .sphinx_gallery_scrapers import *
-from .sphinx_gallery_docs_resolv import *
+from .sphinx_gallery_docs_resolv import _embed_code_links
+from .sphinx_gallery_binder import alt_gen_binder_rst
+from .sphinx_gallery_gen_rst import _save_rst_example
+from .sphinx_gallery_scrapers import alt_matplotlib_scraper
