### 2020-04-07 14091d7
Update Minimum sphinx_gallery version
