### 2020-04-07 c393ae1
Remove redundant docs_resolv patch
