### 2020-04-08 #20 @lwasser
oh awesome!! thank you @rbeucher  and @charlesll  !! i'll add you as reviewers. so here is how it goes

The [review guidelines can be found here](https://www.pyopensci.org/dev_guide/peer_review/reviewer_guide.html). read them over. I need to update that document to contain more recieve pyopensci reviews. when we started we only had ropensci as a model.

When you have completed your review - use the [review template](https://www.pyopensci.org/dev_guide/appendices/templates#Review-Template) so submit your review.

we typically ask for a 3 week turn around on reviews. So if it is possible to **complete your review by Thursday April, 30th** that would be perfect.

At pyopensci we offer the option of including issues that you open as a part of your review that are then addressed through the review. @morganjwilliams has opted to allow this so feel free to submit issues (or pr's) as need be to address things.

Also he has opted to be considered for JOSS. So when we get to that step, we will ask you to have a look at the paper and ensure the package meets the JOSS requirements. we can discuss that as we are further along.

Thank you both for being willing to review this submission. Please get in touch at any point if you have questions OR suggestions. We are still refining our process so we have been tweaking things in each review.
