### 2020-04-09 303c9e0
Add docs pages for pyrolite.data
Update data docs pages
