+++ docs/source/data/aitchison.rst
@@ -1,0 +1,2 @@
+Aitchison Examples
+--------------------------
