+++ docs/source/data/mineral.rst
@@ -1,0 +1,2 @@
+Mineral Database
+------------------------
