+++ docs/source/data/shannon.rst
@@ -1,0 +1,2 @@
+Shannon Radii
+--------------
