+++ docs/source/index.rst
@@ -82,6 +82,7 @@ make use of your geochemical data to build and test geological models.
    :caption: Reference
 
    api/API
+   data/data
    ext/extensions
 
 .. note:: This documentation is a work in progress and is updated regularly. Contact
