+++ docs/source/data/timescale.rst
@@ -1,0 +1,2 @@
+Geological Timescale
+--------------------------
