+++ docs/source/data/data.rst
@@ -1,0 +1,12 @@
+Data
+=====
+
+  .. toctree::
+    :glob:
+    :maxdepth: 1
+
+    refcomp
+    shannon
+    timescale
+    mineral
+    aitchison
