### 2020-04-23 d9a2de5
Update util.lambdas and naming
