+++ pyrolite/util/lambdas.py
@@ -1,16 +1,19 @@
 import numpy as np
+import pandas as pd
 from sympy.solvers.solvers import nsolve
 from sympy import symbols, var
 from functools import partial
 import scipy
+from ..geochem.ind import REE, get_ionic_radii
 from .meta import update_docstring_references
 import logging
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 logger = logging.getLogger(__name__)
 
+
 @update_docstring_references
-def OP_constants(xs, degree=3, rounding=None, tol=10 ** -14):
+def orthogonal_polynomial_constants(xs, degree=3, rounding=None, tol=10 ** -14):
     r"""
     Finds the parameters
     :math:`(\beta_0), (\gamma_0, \gamma_1), (\delta_0, \delta_1, \delta_2)` etc.
@@ -91,9 +94,13 @@ def OP_constants(xs, degree=3, rounding=None, tol=10 ** -14):
     return params
 
 
-def lambda_poly(x, ps):
+########################################################################################
+
+
+def evaluate_lambda_poly(x, ps):
     """
-    Polynomial lambda_n(x) given parameters ps with len(ps) = n.
+    Evaluate polynomial `lambda_n(x)` given a tuple of parameters `ps` with length
+    equal to the polynomial degree.
 
     Parameters
     -----------
@@ -114,112 +121,7 @@ def lambda_poly(x, ps):
     return result.astype(np.float)
 
 
-def lambda_min_func(ls, ys, arrs, power=2.0):
-    """
-    Cost function for lambda optitmization.
-
-    Parameters
-    ------------
-    ls : :class:`numpy.ndarray`
-        Lambda values, effectively weights for the polynomial components.
-    ys : :class:`numpy.ndarray`
-        Target y values.
-    arrs : :class:`numpy.ndarray`
-        Arrays representing the individual unweighted orthaogonal polynomial components.
-        E.g. arrs[0] = `[a, a, a]`, arrs[1] = `[(x-b), (x-b), (x-b)]` etc.
-    power : :class:`float`
-        Power for the cost function. 1 for MAE/L1 norm, 2 for MSD/L2 norm.
-
-    Returns
-    -------
-    :class:`numpy.ndarray`
-        Cost at the given set of `ls`.
-
-    Todo
-    -----
-        * Rewrite cost function for readability with :func:`lambdas`.
-    """
-    cost = np.abs(np.dot(ls, arrs) - ys) ** power
-    cost[np.isnan(cost)] = 0.0  # can't change nans - don't penalise them
-    return cost
-
-
-@update_docstring_references
-def lambdas(
-    arr: np.ndarray,
-    xs=np.array([]),
-    params=None,
-    guess=None,
-    degree=5,
-    costf_power=2.0,
-    residuals=False,
-    min_func=lambda_min_func,
-):
-    """
-    Parameterises values based on linear combination of orthogonal polynomials
-    over a given set of values for independent variable `x`. [#ref_1]_
-
-    Parameters
-    -----------
-    arr : :class:`numpy.ndarray`
-        Target data to fit. For geochemical data, this is typically normalised
-        so we can fit a smooth function.
-    xs : :class:`numpy.ndarray`
-        Values of `x` to construct the polymomials over.
-    params : :class:`list`, :code:`None`
-        Orthogonal polynomial coefficients (see :func:`OP_constants`). Defaults to
-        `None`, in which case these coefficinets are generated automatically.
-    guess : :class:`numpy.ndarray`
-        Starting for values of lambdas. Used as starting point for optimization.
-    degree : :class:`int`
-        Maximum degree polymomial component to include.
-    costf_power : :class:`float`
-        Power of the optimization cost function.
-    residuals : :class:`bool`
[…]
