### 2020-04-23 523bca0
Remove cost_function_power from lambdas function
