### 2020-04-23 47c7619
Update lambdas docstring
