### 2020-04-23 0fe9941
Update _lambda_min_func parameters
