### 2020-04-24 d07c2e0
Update util.lambdas evaluation
