### 2020-04-24 746b9ab
Update geochem.transform and util.lambdas
