### 2020-04-24 c3d023a
Modify util.lambdas test
