### 2020-04-24 7bb0818
Update lambdas example
