### 2020-04-24 d9a23f7
Switch x-axis for lambdas example
Minor docs update
