### 2020-04-28 e231907
Update Travis config to use bionic (Ubuntu 18.x)
