### 2020-04-29 2f7d014
Update lambdas, add ChondriteREE_ON composition
Add the chondrite REE composition from O'Neill 2016 for consistency
