### 2020-04-29 c227d3a
Update pyrolite plot to use df.reindex (#31)
