### 2020-04-29 07d5f2f
Update cross-links to lambdas example
