+++ docs/source/dev/changelog.rst
@@ -6,32 +6,19 @@ All notable changes to this project will be documented here.
 Todo
 ------
 
-* FEATURE: Updates to include more lithogeochemical plot templates
+* **Feature**: Docs page for reference compositions
+  (`#38 <https://github.com/morganjwilliams/pyrolite/issues/38>`__).
+  Work in progress on a
+  `feature branch <https://github.com/morganjwilliams/pyrolite/tree/feature/docs-pyrolite.data>`__.
+* **Feature**: REY function
+  (`#35 <https://github.com/morganjwilliams/pyrolite/issues/35>`__).
+* **Feature**: Updates to include more lithogeochemical plot templates
   (`#26 <https://github.com/morganjwilliams/pyrolite/issues/26>`__)
-* BUG: Conditional density spider plots should have bins centred on the element indexes
+* **Bug**: Conditional density spider plots should have bins centred on the element indexes
   (currently this is an edge)
-* BUG: Index memory for :func:`~pyrolite.plot.spider.spider`
+* **Feature**: Index memory for :func:`~pyrolite.plot.spider.spider`
   (`#27 <https://github.com/morganjwilliams/pyrolite/issues/27>`__)
 
-
-Feature: `convert_chemistry`_ Revamp
-------------------------------------
-
-* BUG: Upgrades for :func:`~pyrolite.geochem.transform.convert_chemistry` for performance,
-  reducing data duplication (`#29 <https://github.com/morganjwilliams/pyrolite/issues/29>`__)
-
-This bug appears to have resulted from caching the function calls to
-:func:`pyrolite.geochem.ind.simple_oxides`, which is addressed with
-`18fede0 <https://github.com/morganjwilliams/pyrolite/commit/18fede01d54d06edd3fe1451409880d889e7ee62>`__.
-
-This branch will remain active to allow a few upgrades to be completed for
-:mod:`pyrolite.geochem.transform` before being merged into `develop`.
-
-Feature: `lambdas`_ Revamp
----------------------------
-
-* BUG: pyrolite lambdas differ slightly from O'Neill (2016) paper
-  (`#39 <https://github.com/morganjwilliams/pyrolite/issues/39>`__)
 
 `Development`_
 --------------
@@ -40,6 +27,10 @@ Feature: `lambdas`_ Revamp
         If you're keen to check something out before its released, you can use a
         `development install <development.html#development-installation>`__.
 
+`0.2.6`_ (Next Release)
+------------------------
+
+* **New Contributors**: `Kaarel Mand <https://github.com/kaarelmand>`__ and Laura Miller
 * Added `codacy` for code quality checking, and implemented numerous clean-ups
   and a few new tests across the package.
 * Performance upgrades, largely for the documentation page.
@@ -47,18 +38,44 @@ Feature: `lambdas`_ Revamp
   due to smaller default image sizes/DPI.
 * Removed dependency on :mod:`fancyimpute`, instead using functions from
   :mod:`scikit-learn`
-* Added the mineral database to `MANIFEST.in` to allow this to be installed with
-  :mod:`pyrolite` (fixing a bug where this isn't present).
 
 :mod:`pyrolite.geochem`
 ~~~~~~~~~~~~~~~~~~~~~~~
 
+* **Bugfix**: pyrolite lambdas differ slightly from [ONeill2016]_
+  (`#39 <https://github.com/morganjwilliams/pyrolite/issues/39>`__). Laura Miller
+  identified inconsistencies between the pyrolite implementation of the lambdas
[…]
