### 2020-04-29 #20 @rbeucher
Hi, please find my review below:

*rbeucher:* This is clearly written in the readthedocs documentation. Please copy and paste to the README

- [X] **Function Documentation:** for all user-facing functions
- [X] **Examples** for all user-facing functions
- [X] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [X] Badges for continuous integration and test coverage, the badge for pyOpenSci peer-review once it has started (see below), a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges, see this example, that one and that one. Such a table should be more wide than high.
- [ ] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.

*rbeucher:* Again, this can just be copied from the documentation.

- [X] Any additional setup required (authentication tokens, etc)
- [ ] Brief demonstration usage

*rbeucher:* The package as a large range of application, I am not sure how this can be done. Maybe a couple of examples.

- [X] Direction to more detailed documentation (e.g. your documentation files or website).
- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages

*rbeucher:* There are probably many packages available on Github that do similar things. Pyrolite as broader applications.

*rbeucher:* 1 test failed: test/plot/plot_spider.py:32

- [X] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

#### For packages co-submitting to JOSS

Pyrolite is without a doubt a very useful package and I will certainly recommend it to my students and colleague.
I have been impressed by the quality of the package. There has been a lot of work in making sure that the functionalities are well documented, both in the docstrings and on readthedocs. The examples provided covers an impressive range of very useful applications. I only had a quick look at the code but it definitely follows some very good practices and I will likely borrow some ideas.

The package is likely to be difficult to use for python beginners but I think that the author has striken the right balance between flexibility and usage. The pandas and matplolib libraries have been leveraged in a way that will provide future-proof development of the code. As a developer myself, I really appreciate the effort and the quality of the work delivered.

There is an extensive range of tests. All of them except 1 have passed on my local machine.
I have spent some time playing with the Jupyter notebook provided on readthedocs.
Some markdown could be added to the notebooks but users can largely figure out what is going on thanks to the docstrings available for each function.
There is a lot of very interesting functionalities (Unit scaling, Convert to oxyde, reference models, composition etc.).

A few Jupyter notebooks break because of a pandas deprecation. This has already been flagged in the issue tracker #31. This must be fixed before accepting the package.

I have also found a few problems due to some data files missing in the data folder. The files were not listed in the MANIFEST.in. I believe this has been fixed in the dev branch and will have to be merged in the version submitted to PyOpenSci.

I would encourage the author to highlight the package as an excellent ressource for teaching in the README.

Having the option to run the examples with Binder is great. However, I found the binders very long to start and I give up at some point. I have not looked at the binder set up but it would be good to improve that if possible.

I hope you will find that review useful.

Romain Beucher
