### 2020-04-29 #20 @morganjwilliams
@rbeucher thanks for your effort in this review and highlighting a few issues! I'll put out a new release shortly (it will be 0.2.6, incorporating changes currently listed in the [changelog 'Development' section](https://pyrolite.readthedocs.io/en/develop/dev/changelog.html#development)) with the majority of these issues addressed and an updated README, and make a note here when it's available.

I've struggled in getting binder to boot up quickly, which is partly due to the relatively heavy dependencies (numpy, pandas, matplotlib, scipy, sklearn), and partly due to the fact that it's likely updated more often than people access the binder links (hence the docker image is rebuilt on-access). Happy to incorporate suggestions to make this easier to work with if anyone has ideas on this front.
