+++ docs/source/dev/contributors.rst
@@ -2,7 +2,7 @@ Contributors
 ============
 
 This list includes people who have contributed to the project in the form of code,
-comments, testing, bug reports, feature requests.
+comments, testing, bug reports, or feature requests.
 
 * `Morgan Williams <https://github.com/morganjwilliams>`__
 * Hayden Dalton
