### 2020-04-29 ffbb676
Update changelog, contributors
