+++ docs/source/dev/contributors.rst
@@ -11,4 +11,4 @@ comments, testing, bug reports, or feature requests.
 * Yajing Mao
 * `Justin Gosses <https://github.com/JustinGOSSES>`__
 * `Kaarel Mand <https://github.com/kaarelmand>`__
-* Laura Miller
+* `Laura Miller <https://github.com/Lauraanme>`__
