### 2020-04-30 53203b3
Add education note to Readme
