### 2020-04-30 8d8908a
Minor formatting updates for util.skl.transform
Update util.skl.transform docstrings


Docstring formatting
