+++ pyrolite/util/skl/transform.py
@@ -28,11 +28,8 @@ except ImportError:
 
 
 class DropBelowZero(BaseEstimator, TransformerMixin):
-    """
-    Transformer for scikit-learn like use.
-    """
-
     def __init__(self, **kwargs):
+        """Transformer for scikit-learn like use."""
         self.kpairs = kwargs
         self.label = "Feedthrough"
 
@@ -48,11 +45,8 @@ class DropBelowZero(BaseEstimator, TransformerMixin):
 
 
 class LinearTransform(BaseEstimator, TransformerMixin):
-    """
-    Linear Transformer for scikit-learn like use.
-    """
-
     def __init__(self, **kwargs):
+        """Linear Transformer for scikit-learn like use."""
         self.kpairs = kwargs
         self.label = "Feedthrough"
         self.forward = lambda x: x
@@ -85,11 +79,8 @@ class LinearTransform(BaseEstimator, TransformerMixin):
 
 
 class ExpTransform(BaseEstimator, TransformerMixin):
-    """
-    Exponential Transformer for scikit-learn like use.
-    """
-
     def __init__(self, **kwargs):
+        """Exponential Transformer for scikit-learn like use."""
         self.kpairs = kwargs
         self.label = "Feedthrough"
         self.forward = np.exp
@@ -118,11 +109,8 @@ class ExpTransform(BaseEstimator, TransformerMixin):
 
 
 class LogTransform(BaseEstimator, TransformerMixin):
-    """
-    Log Transformer for scikit-learn like use.
-    """
-
     def __init__(self, **kwargs):
+        """Log Transformer for scikit-learn like use."""
         self.kpairs = kwargs
         self.label = "Feedthrough"
         self.forward = np.log
@@ -155,11 +143,8 @@ class LogTransform(BaseEstimator, TransformerMixin):
 
 
 class ALRTransform(BaseEstimator, TransformerMixin):
-    """
-    Additive Log Ratio Transformer for scikit-learn like use.
-    """
-
     def __init__(self, **kwargs):
+        """Additive Log Ratio Transformer for scikit-learn like use."""
         self.kpairs = kwargs
         self.label = "ALR"
         self.forward = alr
@@ -192,11 +177,8 @@ class ALRTransform(BaseEstimator, TransformerMixin):
 
 
 class CLRTransform(BaseEstimator, TransformerMixin):
-    """
-    Centred Log Ratio Transformer for scikit-learn like use.
-    """
-
     def __init__(self, **kwargs):
+        """Centred Log Ratio Transformer for scikit-learn like use."""
         self.kpairs = kwargs
         self.label = "CLR"
         self.forward = clr
@@ -229,11 +211,8 @@ class CLRTransform(BaseEstimator, TransformerMixin):
 
 
 class ILRTransform(BaseEstimator, TransformerMixin):
-    """
-    Isometric Log Ratio Transformer for scikit-learn like use.
-    """
-
     def __init__(self, **kwargs):
+        """Isometric Log Ratio Transformer for scikit-learn like use."""
         self.kpairs = kwargs
         self.label = "ILR"
         self.forward = ilr
@@ -255,7 +234,7 @@ class ILRTransform(BaseEstimator, TransformerMixin):
 
     def inverse_transform(self, Y, *args, **kwargs):
         if "X" not in kwargs:
-            if not self.X is not None:
+            if self.X is not None:
                 kwargs.update(dict(X=self.X))
         if isinstance(Y, pd.DataFrame):
             out = pd.DataFrame(
@@ -272,11 +251,8 @@ class ILRTransform(BaseEstimator, TransformerMixin):
 
 
 class BoxCoxTransform(BaseEstimator, TransformerMixin):
-    """
[…]
