### 2020-04-30 f009f15
Minor update for util.lambdas to reduce redundancy
