### 2020-04-30 832771c
Minor cleanup for util.synthetic
