### 2020-04-30 315f9fe
Small cleanup for __init__
