+++ pyrolite/__init__.py
@@ -1,7 +1,11 @@
 """
 pyrolite: A set of tools for getting the most from your geochemical data.
 """
-import sys
+from ._version import get_versions
+
+__version__ = get_versions()["version"]
+del get_versions
+
 import logging
 import importlib
 import pkgutil
@@ -10,11 +14,6 @@ import pkgutil
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 logging.captureWarnings(True)
 
-from ._version import get_versions
-
-__version__ = get_versions()["version"]
-del get_versions
-
 
 def load_extensions(base="pyrolite_", replace=["util"]):
     """
@@ -22,7 +21,7 @@ def load_extensions(base="pyrolite_", replace=["util"]):
     to be importable from :mod:`pyrolite.extensions`.
 
     Parameters
-    ------------
+    ----------
     base : :class:`str`
         Module base string pattern for recognising extensions.
     replace : :class:`list`
@@ -35,7 +34,6 @@ def load_extensions(base="pyrolite_", replace=["util"]):
         for finder, name, ispkg in pkgutil.iter_modules()
         if name.startswith(base)
     }
-    output = {}
     for n, m in modules.items():
         for r in replace:
             n = n.replace(r, "")
