### 2020-04-30 8ca444b
Cleanup geochem.__init__
Update plot.spider use of geochem.ind
