+++ pyrolite/geochem/__init__.py
@@ -14,8 +14,6 @@ from . import parse
 from . import transform
 from . import norm
 from .ind import (
-    common_elements,
-    common_oxides,
     __common_elements__,
     __common_oxides__,
     REE,
@@ -28,11 +26,8 @@ set_default_ionic_charges()
 @pd.api.extensions.register_series_accessor("pyrochem")
 @pd.api.extensions.register_dataframe_accessor("pyrochem")
 class pyrochem(object):
-    """
-    Custom dataframe accessor for pyrolite geochemistry.
-    """
-
     def __init__(self, obj):
+        """Custom dataframe accessor for pyrolite geochemistry."""
         self._validate(obj)
         self._obj = obj
 
@@ -461,7 +456,7 @@ class pyrochem(object):
         radii vs. ln(REE/NORM) polynomial combination.
 
         Parameters
-        ------------
+        ----------
         norm_to : :class:`str` | :class:`~pyrolite.geochem.norm.Composition` | :class:`numpy.ndarray`
             Which reservoir to normalise REE data to (defaults to :code:`"ChondriteREE_ON"`).
         exclude : :class:`list`, :code:`["Pm", "Eu"]`
@@ -475,13 +470,13 @@ class pyrochem(object):
             Current units for the REE data, used to scale the reference dataset.
 
         References
-        -----------
+        ----------
         .. [#localref_1] O’Neill HSC (2016) The Smoothness and Shapes of Chondrite-normalized
                Rare Earth Element Patterns in Basalts. J Petrology 57:1463–1508.
                doi: `10.1093/petrology/egw047 <https://dx.doi.org/10.1093/petrology/egw047>`__
 
         See Also
-        ---------
+        --------
         :func:`~pyrolite.geochem.ind.get_ionic_radii`
         :func:`~pyrolite.util.lambdas.calc_lambdas`
         :func:`~pyrolite.util.lambdas.orthogonal_polynomial_constants`
@@ -502,7 +497,7 @@ class pyrochem(object):
         Attempts to convert a dataframe with one set of components to another.
 
         Parameters
-        -----------
+        ----------
         to : :class:`list`
             Set of columns to try to extract from the dataframe.
 
@@ -517,12 +512,12 @@ class pyrochem(object):
             Flag that data is in molecular units, rather than weight units.
 
         Returns
-        --------
+        -------
         :class:`pandas.DataFrame`
             Dataframe with converted chemistry.
 
         Todo
-        ------
+        ----
             * Check for conflicts between oxides and elements
             * Aggregator for ratios
             * Implement generalised redox transformation.
@@ -539,7 +534,7 @@ class pyrochem(object):
         Normalise a dataframe to a given reference composition.
 
         Parameters
-        -----------
+        ----------
         reference : :class:`str` | :class:`~pyrolite.geochem.norm.Composition` | :class:`numpy.ndarray`
             Reference composition to normalise to.
         units : :class:`str`
@@ -550,12 +545,12 @@ class pyrochem(object):
             Ti, TiO2).
 
         Returns
-        --------
+        -------
         :class:`pandas.DataFrame`
             Dataframe with normalised chemistry.
 
         Notes
-        ------
+        -----
         This assumes that dataframes have a single set of units.
         """
 
@@ -581,19 +576,19 @@ class pyrochem(object):
         De-normalise a dataframe from a given reference composition.
 
         Parameters
-        -----------
+        ----------
         reference : :class:`str` | :class:`~pyrolite.geochem.norm.Composition` | :class:`numpy.ndarray`
             Reference composition which the composition is normalised to.
         units : :class:`str`
             Units of the input dataframe, to convert the reference composition.
 
         Returns
-        --------
+        -------
         :class:`pandas.DataFrame`
[…]
