### 2020-04-30 d353817
Minor formatting for util.skl.impute
