### 2020-04-30 #20 @charlesll
Charles Le Losq: this needs to be added. The website is good but people also stop on the README when browsing Github. I would suggest adding things in the README on Github like a short description of what this software does, its goals, installation, roadmap, etc.

Only for the stable version, not present for the development version.

This is present on the website.

- [x] **Function Documentation:** for all user-facing functions
This is well documented on the website.

- [x] **Examples** for all user-facing functions
Again, very good, well documented on the website.

I would suggest duplicating the contributing guidelines from the website in the CONTRIBUTING.md file on Github.

- [x] **Metadata** including author(s), author e-mail(s), a url, and any other relevant metadata e.g., in a `setup.py` file or elsewhere.

Readme requirements

- [x] Badges for continuous integration and test coverage, the badge for pyOpenSci peer-review once it has started (see below), a repostatus.org badge, and any other badges. If the README has many more badges, you might want to consider using a table for badges, see this example, that one and that one. Such a table should be more wide than high.
- [x] Short description of goals of package, with descriptive links to all vignettes (rendered, i.e. readable, cf the documentation website section) unless the package is small and there’s only one vignette repeating the README.
- [x] Any additional setup required (authentication tokens, etc)
- [ ] Brief demonstration usage
Not present. This can be added.
- [x] Direction to more detailed documentation (e.g. your documentation files or website).
- [ ] If applicable, how the package compares to other similar packages and/or how it relates to other packages
I think it needs to be added on the Github README. Present on the website.

I would suggest integrating the tests through nosetests. See https://python-packaging.readthedocs.io/en/latest/testing.html

- [x] **Continuous Integration:** Has continuous integration, such as Travis CI, AppVeyor, CircleCI, and/or others.

#### For packages co-submitting to JOSS

While the package mostly proposes utility functions, its global use can definitely allow developing geochemical models. Therefore, I think it fits JOSS guidelines.

Estimated hours spent reviewing: 3 h

# Revier Comments

Pyrolite is a useful package in a community lead by Excel or other commercial software.

It will help the manipulation of geochemical data. It is not limited to this discipline, as actually it
can help the manipulation, visualization and interpretation of data where ternary plots, chemical conversion and
so on are necessary. The plans to link it to python-MELTS, to develop isotopic calculation and so on are also very good news.
I particularly encourage the author to make the effort to link it to python-MELTS, as this can be
an important source of users.

Some areas of improvement include:

- As a general advice, I just wanted to say that I encourage the author to continue to provide clear examples of how the plotting API inferface with Matplotlib in the future. Matplotlib is THE standard Python plotting library. You should aim at complementing it, not replacing it.
In general, when one proposes an easier way to make some plots difficult to do with Matplotlib, it is easier at first, but then I want to
do maybe something more complex and at this stage, usually, I have to re-write everything in Matplotlib... This is why I advice you to make sure that your
plotting functions interface seamlessly with Matplotlib. Please keep that in mind upon further developing your package.

- I would suggest to add the example notebooks in an example folder in your root directory in Github. Otherwise it is actually difficult to find them. (I like to see the notebooks online on Github, it is very boring to have to download them...) Also, a bit more details and comments in the notebooks will be better. I personally prefer documented Jupyter notebooks compared to websites.

- Maybe you would like to add an option in get_ionic_radii() for using the Whittaker and Muntus (1970) ionic radii that were maybe more geared toward geochemical oxides ?

- Add an example of how to properly put subscribs for indicating SiO2, or Al2O3 or any other element with subscripts in ternary plots. Without that, the ternary plots are not really "publishable" as is...
