### 2020-05-01 morganjwilliams/pyrolite#41 @morganjwilliams
Add Whittaker and Muntus (1970) Ionic Radii
In addition to the Shannon radii, we could add an option to [`pyrolite.geochem.ind.get_ionic_radii`](https://pyrolite.readthedocs.io/en/develop/api/geochem/ind.html#pyrolite.geochem.ind.get_ionic_radii) to return the Whittaker and Muntus radii, which are likely more suitable for silicate geochemistry.

**Relevant References**

Whittaker, E.J.W., Muntus, R. (1970). Ionic radii for use in geochemistry. Geochimica et Cosmochimica Acta 34, 945–956. doi: [10.1016/0016-7037(70)90077-3](https://doi.org/10.1016/0016-7037(70)90077-3)
