### 2020-05-01 morganjwilliams/pyrolite#42 @morganjwilliams
Add Tutorial for Formatting Plot Labels, Titles etc
The documentation doesn't currently have much information about formatting and adjusting plots which are created. While the pyrolite docs shouldn't directly replicate those for matplotlib, there's a good argument for including a tutorial for these aspects of creating plots.

In particular, things like adjusting formatting for ternary diagrams (which uses some [mpltern](https://mpltern.readthedocs.io/en/latest/)-specific syntax), using subscripts and superscripts, incorporating TeX, and plotting over your existing figures (or how to integrate pyrolite with other matplotlib figures etc) would all be useful inclusions.
