### 2020-05-01 #20 @morganjwilliams
Thanks @charlesll! I'll put together a to-do list of things to update and add a note here when 0.2.6 is out, as mentioned above. I've added two issues to address some of your later comments ([#41](https://github.com/morganjwilliams/pyrolite/issues/41) & [#42](https://github.com/morganjwilliams/pyrolite/issues/42)).

On the MELTS front, there's an incomplete wrapper for the alphaMELTS executable at [pyrolite-meltsutil.readthedocs.io](https://pyrolite-meltsutil.readthedocs.io/en/develop/), but I'm looking forward to working with the newer Python-MELTS API when it's ready to go.

In general most aspects of the matplotlib API used in pyrolite are exposed, and relevant arguments are passed through to the matplotlib (or in the case of ternary plots, mpltern) functions and classes to make sure it integrates well. Necessarily a few things are lost in translation, but let me know if anything in particular is concerning here.

As a comment on the general submission requirements, I'm curious how much duplication is advised to fill out the README etc, just because it might end up being both quite lengthy and slightly more difficult to maintain/keep up to date.
