### 2020-05-05 #20 @charlesll
@morganjwilliams  for the README, just add a short version that will be understandable by people already used to Python / coding. But I think it is necessary as those people usually search on Github, and the README is the first thin that appears.

Regarding Matplotlib, all seems fine for now, my comment was mosty to invite you to keep it that way. For instance, I like mpltern because it is a very thin wrapper around Matplotlib to do ternary plots. Personally I know how to use pyplot and it does the job, and I won't invest time in learning another plotting package because I just don't have the time. Even when coding in Julia I use Pyplot. I guess I am not the only one in this situation. This is why if you keep your plotting stuffs close to the MPL API, I think this is best and that it will help the growth of your user-base.
