### 2020-05-06 #20 @morganjwilliams
@charlesll no worries, can do. I feel the struggle about vis libraries, although I've thought about adding an interactive plotting backend option for pyrolite via e.g. [hvplot](https://hvplot.holoviz.org/index.html) (but that's one for the future, and I need to look at feasibility more).  Expect 0.2.6 to be out at the end of the week.
