### 2020-05-06 c5da030
Update docs/conf.py to use updated paths
