### 2020-05-06 ed00203
Bugfix for breaking changes in TinyDB v4.0.0
https://tinydb.readthedocs.io/en/latest/changelog.html#v4-0-0-2020-05-02
Typo for table.truncate (tinydb fix)
