### 2020-05-06 b6d7193
Update db table reference in geochem.norm
Update comment about db.table
