+++ pyrolite/geochem/ind.py
@@ -19,10 +19,22 @@ import logging
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 logger = logging.getLogger(__name__)
 
-_shannonradiifile = (pyrolite_datafolder(subfolder="shannon") / "radii.csv").resolve()
-assert _shannonradiifile.exists() and _shannonradiifile.is_file()
-__shannon__ = pd.read_csv(_shannonradiifile).set_index("index", drop=True)
-assert hasattr(__shannon__, "element")
+__radii__ = {}
+
+
+def _load_radii():
+    """Import radii tables to a module-level dictionary indexed by reference."""
+    global __radii__
+    for name in ["shannon", "whittaker_muntus"]:
+        pth = (pyrolite_datafolder(subfolder="radii") / "{}.csv".format(name)).resolve()
+        assert pth.exists() and pth.is_file()
+        df = pd.read_csv(pth).set_index("index", drop=True)
+        assert hasattr(df, "element")
+        __radii__[name] = df
+
+
+_load_radii()
+########################################################################################
 
 
 @functools.lru_cache(maxsize=4)  # cache outputs for speed
@@ -168,7 +180,6 @@ def common_oxides(
         return oxides
 
 
-#@functools.lru_cache(maxsize=None)  # cache outputs for speed
 def simple_oxides(cation, output="string"):
     """
     Creates a list of oxides for a cationic element (oxide of ions with c=1+ and above).
@@ -263,98 +274,6 @@ def get_isotopes(ratio_text):
         return bw
 
 
-def get_ionic_radii(element, charge=None, coordination=None, variant=[], pauling=True):
-    """
-    Function to obtain Shannon's radii for a given ion [1]_. Shannon published two sets of
-    radii. The first ('Crystal Radii') were using Shannon's value for :math:`r(O^{2-}_{VI})`
-    of 1.26 Å, while the second ('Ionic Radii') is consistent with the
-    Pauling (1960) value of :math:`r(O^{2-}_{VI})` of 1.40 Å [2]_.
-
-    Parameters
-    -----------
-    element : :class:`str` | :class:`list`
-        Element to obtain a radii for. If a list is passed, the function will be applied
-        over each of the items.
-    charge : :class:`int`
-        Charge of the ion to obtain a radii for. If unspecified will use the default
-        charge from :mod:`pyrolite.mineral.ions`.
-    coordination : :class:`int`
-        Coordination of the ion to obtain a radii for.
-    variant : :class:`list`
-        List of strings specifying particular variants (here 'squareplanar' or
-        'pyramidal', 'highspin' or 'lowspin').
-    pauling : :class:`bool`
-        Whether to use the radii consistent with Pauling (1960).
-
-    Returns
-    --------
-    :class:`pandas.Series`
-        Dataframe with viable ion charge and coordination, with associated radii in
-        angstroms. If the ion charge and coordiation are specified and found in the
-        table, a single value will be returned instead.
-
-    References
-    ----------
-    .. [1] Shannon RD (1976). Revised effective ionic radii and systematic
-            studies of interatomic distances in halides and chalcogenides.
-            Acta Crystallographica Section A 32:751–767.
-            doi: shannon1976
-    .. [2] Pauling, L., 1960. The Nature of the Chemical Bond.
-            Cornell University Press, Ithaca, NY.
-
-    Todo
-    -----
-    * Implement interpolation for coordination +/- charge.
-    * Finish Shannon Radii tests
-    """
-    if isinstance(element, list):
-        return [
-            get_ionic_radii(
-                e,
-                charge=charge,
-                coordination=coordination,
-                variant=variant,
-                pauling=pauling,
-            )
-            for e in element
-        ]
-
[…]
