### 2020-05-06 dfc511a
Update radii function references for geochem.ind
