### 2020-05-06 f625942
Update util.meta test for data folders
