### 2020-05-06 44aa03f
Update docstring for geochem.ind.get_ionic_radii
