### 2020-05-06 morganjwilliams/pyrolite#41 @morganjwilliams
This table of radii was added, and the feature branch merged into develop with 9620b817747009c4d4262af41d06f65d7fcb4d92. See the API docs for [pyrolite.geochem.ind.get_ionic_radii](https://pyrolite.readthedocs.io/en/develop/api/geochem/ind.html#pyrolite.geochem.ind.get_ionic_radii) for more info.

I'll soon update the [relevant example](https://pyrolite.readthedocs.io/en/develop/examples/geochem/shannon.html#sphx-glr-examples-geochem-shannon-py) to show how the optional argument can be used (e.g, `source='whittaker'`) and compare the two sets of ionic radii (or three, given Shannon 1976 has two sets itself).

Thanks to @charlesll for the suggestion!
