### 2020-05-06 9089e6a
Update geochem.ind.get_ionic radii docstring
