### 2020-05-06 a9b8c2f
Ensure kwargs are passed for iteration over lists for get_ionic_radii
