+++ docs/source/examples/geochem/shannon.rst
@@ -1,82 +1,0 @@
-.. rst-class:: sphx-glr-example-title
-
-.. _sphx_glr_examples_geochem_shannon.py:
-
-
-Shannon Radii
-=============
-
-
-.. code-block:: default
-
-    import matplotlib.pyplot as plt
-    from pyrolite.geochem.ind import get_ionic_radii, REE
-
-    REE_radii = get_ionic_radii(REE(), coordination=8, charge=3)
-
-    fig, ax = plt.subplots(1)
-    ax.plot(REE_radii, c='k', marker='D')
-    ax.set_xticks(range(len(REE())))
-    ax.set_xticklabels(REE())
-    ax.set_ylabel('Ionic Radius ($\AA$)')
-    ax.set_title('Rare Earth Element Shannon Ionic Radii')
-
-
-
-.. image:: /examples/geochem/images/sphx_glr_shannon_001.png
-    :class: sphx-glr-single-img
-
-
-.. rst-class:: sphx-glr-script-out
-
- Out:
-
- .. code-block:: none
-
-
-    Text(0.5, 1.0, 'Rare Earth Element Shannon Ionic Radii')
-
-
-
-.. seealso:: `Pandas Lambda Ln(REE) Function <../lambdas/pandaslambdas.html>`__,
-             `Lambdas for Dimensional Reduction <../lambdas/lambdadimreduction.html>`__,
-             `REE Radii Plot <../plotting/REE_radii_plot.html>`__
-
-
-.. rst-class:: sphx-glr-timing
-
-   **Total running time of the script:** ( 0 minutes  0.206 seconds)
-
-
-.. _sphx_glr_download_examples_geochem_shannon.py:
-
-
-.. only :: html
-
- .. container:: sphx-glr-footer
-    :class: sphx-glr-footer-example
-
-
-  .. container:: binder-badge
-
-    .. image:: https://mybinder.org/badge_logo.svg
-      :target: https://mybinder.org/v2/gh/morganjwilliams/pyrolite/develop?filepath=docs/source/examples/geochem/shannon.ipynb
-      :width: 150 px
-
-
-  .. container:: sphx-glr-download sphx-glr-download-python
-
-     :download:`Download Python source code: shannon.py <shannon.py>`
-
-
-
-  .. container:: sphx-glr-download sphx-glr-download-jupyter
-
-     :download:`Download Jupyter notebook: shannon.ipynb <shannon.ipynb>`
-
-
-.. only:: html
-
- .. rst-class:: sphx-glr-signature
-
-    `Gallery generated by Sphinx-Gallery <https://sphinx-gallery.github.io>`_
