### 2020-05-07 0e505f7
Add Plot Formatting Tutorial Page
