+++ docs/source/gallery/tutorials/plot_formatting.py
@@ -1,0 +1,7 @@
+"""
+Formatting and Cleaning Up Plots
+==================================
+
+.. note:: This tutorial is a work in progress and will be gradually updated.
+
+"""
