### 2020-05-07 c53dc6b
Add 'Education' to intended audience in setup.py, link to matplotlib
Add matplotlib to framework
