### 2020-05-07 799eeff
Update util.plot.grid test
Consider refactoring later (with plot.density.grid, plot.density.ternary)
