### 2020-05-07 5763505
Update tests for util.plot.helpers
