### 2020-05-07 fbe9172
Update plot.color testing
