### 2020-05-07 5fcd738
Update tests for util.general
