+++ pyrolite/util/skl/impute.py
@@ -38,12 +38,14 @@ class MultipleImputer(BaseEstimator, TransformerMixin):
         assert isinstance(X, pd.DataFrame)
         df = pd.DataFrame(columns=X.columns, index=X.index)
         if isinstance(self.imputers, dict):
-            for d in self.imputers.values():
-                mask = d["mask"]
-                imputers = d["impute"]
-                imputed_data = np.array([imp.transform(X[mask, :]) for imp in imputers])
+            for cls, content in self.imputers.items():
+                mask = content["mask"]
+                imputers = content["impute"]
+                imputed_data = np.array(
+                    [imp.transform(X.loc[mask, :]) for imp in imputers]
+                )
                 mean = np.mean(imputed_data, axis=0)
-                df.loc[mask, ~pd.isnull(X[mask, :]).all(axis=0)] = mean
+                df.loc[mask, ~pd.isnull(X.loc[mask, :]).all(axis=0)] = mean
             return df
         else:
             imputed_data = np.array([imp.transform(X) for imp in self.imputers])
@@ -51,26 +53,36 @@ class MultipleImputer(BaseEstimator, TransformerMixin):
             df.loc[:, ~pd.isnull(X).all(axis=0)] = mean
             return df
 
-
-    #def inverse_transform(self, Y, *args, **kwargs):
+    # def inverse_transform(self, Y, *args, **kwargs):
     #    # For non-compositional data, take the mask and reverting to nan
     #    # for compositional data, renormalisation would be needed
     #    pass
 
     def fit(self, X, y=None):
+        """
+        Fit the imputers.
+
+        Parameters
+        ----------
+        X : :class:`pandas.DataFrame`
+            Data to use to fit the imputations.
+        y : :class:`pandas.Series`
+            Target class; optionally specified, and used similarly to `groupby`.
+        """
         assert isinstance(X, pd.DataFrame)
         # start = X
         y_present = y is not None
         groupby_present = self.groupby is not None
         self.imputers = []
         if y_present or groupby_present:
+            # here works for one or the other, but could technically split for this
             assert not (groupby_present and y_present)
             if y_present:
                 classes = np.unique(y)
-                gen_mask = lambda c: y == c
+                gen_mask = lambda c: np.array(y == c)
             if groupby_present:
                 classes = X[self.groupby].unique()
-                gen_mask = lambda c: X[self.groupby] == c
+                gen_mask = lambda c: np.array(X[self.groupby] == c)  # pd.Series values
             self.imputers = {
                 c: {
                     "impute": [
@@ -92,9 +104,9 @@ class MultipleImputer(BaseEstimator, TransformerMixin):
             )
             logger.info(msg)
 
-            for d in self.imputers.values():
-                for imp in d["impute"]:
-                    imp.fit(X[d["mask"], :])
+            for cls, content in self.imputers.items():
+                for imp in content["impute"]:
+                    imp.fit(X.loc[content["mask"], :])
 
         else:
             for ix in range(self.multiple):
