### 2020-05-08 e9b119d
Update test for util.plot.density
Will need to update percentile_contour_values_from_meshz to deal with arrays of percentiles better
