### 2020-05-08 b11309d
Add statsmodels to travis install
