### 2020-05-08 d4f8631
Update plot.density contours test
Only available here where mode='density', but could re-jig using the integration method to work for others.
