### 2020-05-08 2d746e5
Update geochem.ind, util.lambdas tests
