### 2020-05-08 7671dc6
Move plot_formatting.py, update heading
