+++ docs/source/gallery/tutorials/plot_formatting.py
@@ -4,4 +4,214 @@ Formatting and Cleaning Up Plots
 
 .. note:: This tutorial is a work in progress and will be gradually updated.
 
+In this tutorial we will illustrate some straightfoward formatting for your plots which
+will allow for greater customisation as needed. As :mod:`pyrolite` heavily uses
+and exposes the API of :mod:`matplotlib` for the visualisation components
+(and also :mod:`mpltern` for ternary diagrams), you should also check out their
+documentation pages for more in-depth guides, examples and API documentation.
 """
+#######################################################################################
+# First let's pull in a simple dataset to use throughout these examples:
+#
+from pyrolite.util.synthetic import test_df
+
+df = test_df(cols=["SiO2", "CaO", "MgO", "Al2O3", "TiO2", "27Al", "d11B"])
+#######################################################################################
+# Basic Figure and Axes Settings
+# ------------------------------
+#
+# :mod:`matplotlib` makes it relatively straightfoward to customise most settings for
+# your figures and axes. These settings can be defined at creation (e.g. in a call to
+# :func:`~matplotlib.pyplot.subplots`), or they can be defined after you've created an
+# axis (with the methods :code:`ax.set_<parameter>()`). For example:
+#
+import matplotlib.pyplot as plt
+
+fig, ax = plt.subplots(1)
+
+ax.set_xlabel("My X Axis Label")
+ax.set_title("My Axis Title", fontsize=12)
+ax.set_yscale("log")
+ax.set_xlim((0.5, 10))
+
+fig.suptitle("My Figure Title", fontsize=15)
+
+plt.show()
+#######################################################################################
+# You can use a single method to set most of these things:
+# :func:`~matplotlib.axes.Axes.set`. For example:
+#
+import matplotlib.pyplot as plt
+
+fig, ax = plt.subplots(1)
+ax.set(yscale="log", xlim=(0, 1), ylabel="YAxis", xlabel="XAxis")
+plt.show()
+#######################################################################################
+# Labels and Text
+# ----------------
+#
+# :mod:`matplotlib` enables you to use :math:`\TeX` within all text elements, including
+# labels and annotations. This can be leveraged for more complex formatting,
+# incorporating math and symbols into your plots. Check out the mod:`matplotlib`
+# `tutorial <https://matplotlib.org/3.2.1/tutorials/text/mathtext.html>`__, and
+# for more on working with text generally in :mod:`matplotlib`, check out the
+# `relevant tutorials gallery <https://matplotlib.org/3.2.1/tutorials/index.html#text>`__.
+#
+# The ability to use TeX syntax in :mod:`matplotlib` text objects can also be used
+# for typsetting, like for subscripts and superscripts. This is particularly relevant
+# for geochemical oxides labels (e.g. Al2O3, which would ideally be rendered as
+# :math:`Al_2O_3`) and isotopes (e.g. d11B, which should be :math:`\delta^{11}B`).
+# At the moment, pyrolite won't do this for you, so you may want to adjust the labelling
+# after you've made them. For example:
+import pyrolite.plot
+import matplotlib.pyplot as plt
+
+fig, ax = plt.subplots(2, 1)
+df[["Al2O3", "TiO2"]].pyroplot.scatter(ax=ax[0])
+ax[0].set_xlabel("Al$_2$O$_3$")
+ax[0].set_ylabel("TiO$_2$")
+
+df[["27Al", "d11B"]].pyroplot.scatter(ax=ax[1])
+ax[1].set_xlabel("$^{27}$Al")
+ax[1].set_ylabel("$\delta^{11}$B")
+
+plt.tight_layout() # rearrange the plots to fit nicely together
+plt.show()
[…]
