### 2020-05-08 morganjwilliams/pyrolite#42 @morganjwilliams
This tutorial page has now been merged into the feature branch (thanks to @lavender22), and will soon be on develop ready to release for 0.2.6
