+++ docs/source/tutorials/sg_execution_times.rst
@@ -5,10 +5,12 @@
 
 Computation times
 =================
-**00:10.824** total execution time for **tutorials** files:
+**00:04.250** total execution time for **tutorials** files:
 
 +-----------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_tutorials_logo.py` (``logo.py``)                       | 00:07.227 | 0.0 MB |
+| :ref:`sphx_glr_tutorials_plot_formatting.py` (``plot_formatting.py``) | 00:04.250 | 0.0 MB |
 +-----------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_tutorials_ternary_density.py` (``ternary_density.py``) | 00:03.597 | 0.0 MB |
+| :ref:`sphx_glr_tutorials_logo.py` (``logo.py``)                       | 00:00.000 | 0.0 MB |
++-----------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_tutorials_ternary_density.py` (``ternary_density.py``) | 00:00.000 | 0.0 MB |
 +-----------------------------------------------------------------------+-----------+--------+
