### 2020-05-08 e8434a1
Add Development installation instructions to Readme
