### 2020-05-08 10da410
Edit contributing guidelines
Sphinx_gallery links won't work nicely with 'edit on github' links
