### 2020-05-08 a3d15ff
Expand contributing guidelines in readme
