### 2020-05-08 89eafc8
Duplicate contribuitng.rst into Contributing.md
