### 2020-05-08 f16840f
Update changelog for pre-release
