### 2020-05-11 #20 @morganjwilliams
I've now released pyrolite v0.2.6, with an updated and expanded README, a few new features and some bugfixes (see [changelog](https://pyrolite.readthedocs.io/en/develop/dev/changelog.html#id1)).

This release addresses two of the potential upgrades suggested by @charlesll ([pyrolite#41](https://github.com/morganjwilliams/pyrolite/issues/41) and [pyrolite#42](https://github.com/morganjwilliams/pyrolite/issues/42)), the fix for the missing line in MANIFEST.in mentioned by @rbeucher, and a few updates to get test coverage to about 90%.
