### 2020-05-12 77dff62
Update radii data
