+++ MANIFEST.in
@@ -5,7 +5,7 @@ recursive-include pyrolite/data/Aitchison *
 include pyrolite/data/alphamelts/env.py
 recursive-include pyrolite/data/geochem *
 recursive-include pyrolite/data/models *
-recursive-include pyrolite/data/shannon *
+recursive-include pyrolite/data/radii *
 recursive-include pyrolite/data/timescale *
 recursive-include pyrolite/data/mineral *
 # recursive-exclude pyrolite/data/alphamelts/localinstall *
