### 2020-05-12 b3cb152
Reorder MANIFEST.in for clarity
