+++ MANIFEST.in
@@ -2,11 +2,9 @@ include versioneer.py
 include pyrolite/_version.py
 include LICENSE
 recursive-include pyrolite/data/Aitchison *
-include pyrolite/data/alphamelts/env.py
 recursive-include pyrolite/data/geochem *
+recursive-include pyrolite/data/mineral *
 recursive-include pyrolite/data/models *
 recursive-include pyrolite/data/radii *
 recursive-include pyrolite/data/timescale *
-recursive-include pyrolite/data/mineral *
-# recursive-exclude pyrolite/data/alphamelts/localinstall *
 # global-exclude __pycache__
