### 2020-05-18 2060cd8
Update comp.codata.logratiomean
