### 2020-05-18 60b9f97
Update lattice example doi links
