### 2020-05-18 a3c3521
Update example galleries for pyrolite.comp
