+++ docs/source/gallery/examples/comp/logtransforms.py
@@ -1,0 +1,100 @@
+"""
+Log-transforms
+----------------
+
+pyrolite includes a few functions for dealing with compositional data, at the heart of
+which are i) closure (i.e. everything sums to 100%) and ii) log-transforms to deal with
+the compositional space. The commonly used log-transformations include the
+Additive Log-Ratio (:func:`~pyrolite.comp.pyrocomp.ALR`), Centred Log-Ratio
+(:func:`~pyrolite.comp.pyrocomp.CLR`), and Isometric Log-Ratio
+(:func:`~pyrolite.comp.pyrocomp.ILR`) [#ref_1]_ [#ref_2]_.
+
+This example will show you how to access and use some of these functions in pyrolite.
+"""
+
+########################################################################################
+# First let's create some example data:
+#
+from pyrolite.util.synthetic import test_df, random_cov_matrix
+
+df = test_df(
+    index_length=100,
+    cov=random_cov_matrix(sigmas=[0.1, 0.05, 0.3, 0.6], dim=4, seed=32),
+    seed=32,
+)
+df.describe()
+########################################################################################
+# Let's have a look at some of the log-transforms, which can be accessed directly from
+# your dataframes (via :class:`pyrolite.comp.pyrocomp`), after you've imported
+# :mod:`pyrolite.comp`. Note that the transformations will return *new* dataframes,
+# rather than modify their inputs. For example:
+#
+import pyrolite.comp
+
+lr_df = df.pyrocomp.CLR()  # using a centred log-ratio transformation
+########################################################################################
+# The transformations are implemented such that the column names generally make it
+# evident which transformations have been applied:
+#
+lr_df.columns
+########################################################################################
+# To invert these transformations, you can call the respective inverse transform:
+#
+back_transformed = lr_df.pyrocomp.inverse_CLR()
+########################################################################################
+# Given we haven't done anything to our dataframe in the meantime, we should be back
+# where we started, and our values should all be equal within numerical precision.
+# To verify this, we can use :func:`numpy.allclose`:
+#
+import numpy as np
+
+np.allclose(back_transformed, df)
+########################################################################################
+# In addition to easy access to the transforms, there's also a convenience function
+# for taking a log-transformed mean (log-transforming, taking a mean, and inverse log
+# transforming; :func:`~pyrolite.comp.codata.pyrocomp.logratiomean`):
+#
+
+df.pyrocomp.logratiomean()
+########################################################################################
+# While this function defaults to using :func:`~pyrolite.comp.codata.clr`,
+# you can specify other log-transforms to use:
+#
+df.pyrocomp.logratiomean(transform="clr")
+########################################################################################
+# Notably, however, the logratio means should all give you the same result:
+#
+np.allclose(
+    df.pyrocomp.logratiomean(transform="clr"),
+    df.pyrocomp.logratiomean(transform="alr"),
+) & np.allclose(
+    df.pyrocomp.logratiomean(transform="clr"),
+    df.pyrocomp.logratiomean(transform="ilr"),
+)
+
+########################################################################################
+# .. [#ref_1] Aitchison, J., 1984. The statistical analysis of geochemical compositions.
+#       Journal of the International Association for Mathematical Geology 16, 531–564.
+#       doi: `10.1007/BF01029316 <https://doi.org/10.1007/BF01029316>`__
+#
+# .. [#ref_2]  Egozcue, J.J., Pawlowsky-Glahn, V., Mateu-Figueras, G.,
[…]
