### 2020-05-18 2f2e2bd
Update contributing.md formatting
