### 2020-05-18 f125494
Remove old cached log ratio example
