### 2020-05-18 7bb5827
Update pyrolite.comp.pyrocomp to use df.attrs
