### 2020-05-20 3345bd6
Update util API docs
