+++ pyrolite/util/plot/legend.py
@@ -1,5 +1,12 @@
 """
 Functions for creating and modifying legend entries for matplotlib.
+
+Todo
+------
+
+    * Functions for working with and modifying legend entries.
+
+        ax.lines + ax.patches + ax.collections + ax.containers, handle ax.parasites
 """
 import matplotlib.patches
 import matplotlib.lines
