+++ pyrolite/util/plot/__init__.py
@@ -1,14 +1,7 @@
 """
 Utility functions for working with matplotlib.
 
-Todo
-------
-
-    * Functions for working with and modifying legend entries.
-
-        ax.lines + ax.patches + ax.collections + ax.containers, handle ax.parasites
-
-Attributes
+Parameters
 ----------
 DEFAULT_CONT_COLORMAP : :class:`matplotlib.colors.ScalarMappable`
     Default continuous colormap.
@@ -17,7 +10,6 @@ DEFAULT_DICS_COLORMAP : :class:`matplotlib.colors.ScalarMappable`
 USE_PCOLOR : :class:`bool`
     Option to use the :func:`matplotlib.pyplot.pcolor` function in place
     of :func:`matplotlib.pyplot.pcolormesh`.
-
 """
 import logging
 
