+++ docs/source/api/util/skl.rst
@@ -1,33 +1,38 @@
-pyrolite\.util\.skl\.vis
+pyrolite\.util\.skl
 -------------------------------
+  .. automodule:: pyrolite.util.skl
+
+
+pyrolite\.util\.skl\.vis
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   .. automodule:: pyrolite.util.skl.vis
       :members:
       :undoc-members:
 
 
 pyrolite\.util\.skl\.pipeline
--------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   .. automodule:: pyrolite.util.skl.select
       :members:
       :undoc-members:
 
 
 pyrolite\.util\.skl\.select
--------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   .. automodule:: pyrolite.util.skl.select
       :members:
       :undoc-members:
 
 
 pyrolite\.util\.skl\.transform
--------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   .. automodule:: pyrolite.util.skl.transform
       :members:
       :undoc-members:
 
 
 pyrolite\.util\.skl\.impute
--------------------------------
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
   .. automodule:: pyrolite.util.skl.impute
       :members:
       :undoc-members:
