+++ docs/source/api/util/time.rst
@@ -1,0 +1,5 @@
+pyrolite\.util\.time
+-------------------------------
+  .. automodule:: pyrolite.util.time
+      :members:
+      :undoc-members:
