+++ docs/source/api/util/pd.rst
@@ -1,0 +1,5 @@
+pyrolite\.util\.pd
+-------------------------------
+  .. automodule:: pyrolite.util.pd
+      :members:
+      :undoc-members:
