+++ docs/source/api/util.rst
@@ -1,99 +1,0 @@
-pyrolite\.util\.general
--------------------------------
-  .. automodule:: pyrolite.util.general
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.pd
--------------------------------
-  .. automodule:: pyrolite.util.pd
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.plot
--------------------------------
-  .. automodule:: pyrolite.util.plot
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.text
--------------------------------
-  .. automodule:: pyrolite.util.text
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.web
--------------------------------
-  .. automodule:: pyrolite.util.web
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.time
--------------------------------
-  .. automodule:: pyrolite.util.time
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.math
--------------------------------
-  .. automodule:: pyrolite.util.math
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.lambdas
--------------------------------
-  .. automodule:: pyrolite.util.lambdas
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.distributions
--------------------------------
-  .. automodule:: pyrolite.util.distributions
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.synthetic
--------------------------------
-  .. automodule:: pyrolite.util.synthetic
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.missing
--------------------------------
-  .. automodule:: pyrolite.util.missing
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.units
--------------------------------
-  .. automodule:: pyrolite.util.units
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.types
--------------------------------
-  .. automodule:: pyrolite.util.types
-      :members:
-      :undoc-members:
-
-pyrolite\.util\.meta
--------------------------------
-  .. automodule:: pyrolite.util.meta
-      :members:
-      :undoc-members:
-
-
-pyrolite\.util\.skl
--------------------------------
-  .. automodule:: pyrolite.util.skl
-
-  .. toctree::
-     :maxdepth: 2
-
-     skl
-
-pyrolite\.util\.classification
--------------------------------
-  .. automodule:: pyrolite.util.classification
-      :members:
-      :undoc-members:
