+++ docs/source/api/util/meta.rst
@@ -1,0 +1,5 @@
+pyrolite\.util\.meta
+-------------------------------
+  .. automodule:: pyrolite.util.meta
+      :members:
+      :undoc-members:
