+++ docs/source/api/util/units.rst
@@ -1,0 +1,5 @@
+pyrolite\.util\.units
+-------------------------------
+  .. automodule:: pyrolite.util.units
+      :members:
+      :undoc-members:
