+++ docs/source/api/util/classification.rst
@@ -1,0 +1,5 @@
+pyrolite\.util\.classification
+-------------------------------
+  .. automodule:: pyrolite.util.classification
+      :members:
+      :undoc-members:
