+++ docs/source/api/util/web.rst
@@ -1,0 +1,5 @@
+pyrolite\.util\.web
+-------------------------------
+  .. automodule:: pyrolite.util.web
+      :members:
+      :undoc-members:
