### 2020-05-22 #20 @lwasser
👋 everyone!! checking in on this review. it sounds like @morganjwilliams has implemented a new release. @charlesll  @rbeucher are you satisfied with the fixes as implemented? is there anything else that we need to check or do to finish up this review. thank you all!
just a note that i'll be out of the office until june 1. so i will be unresponsive next week but plan to check back in on June 1 to see if we can wrap this review up.

@morganjwilliams i see that you are interested in submitting this to JOSS. is that still the case? if so you will need to create a paper.md file following JOSS specifications. i see that here: https://github.com/morganjwilliams/pyrolite/tree/master/paper @charlesll @rbeucher did you have a look at that paper as well? many thanks all. once we have ok's on both the review, new release AND the paper we can pass it on to JOSS. the process with JOSS is quick!!
