### 2020-05-22 #20 @rbeucher
Hi @lwasser .

All good I think.
I had a look a the paper. It is ready for JOSS.
