### 2020-05-25 #20 @morganjwilliams
@lwasser Yes, still interested in submitting this to JOSS. Please forward this on once @charlesll has a chance to respond to confirm that he's also had a look at the paper.md and the issues raised above have been addressed.
