### 2020-05-28 3895890
Bugfix for geochem.transform.lambda_lnREE
