### 2020-05-28 9316721
Bugfix for plot.color processing alpha
