### 2020-05-28 493fcd4
Add bugfix notes to changelog
