### 2020-05-29 f3d0d69
Enable passing of bw_method to scipy.stats.guassian_kde for density
