### 2020-06-02 5afe7c2
Bugfix for util.lambdas
Fixed issue with handling of missing values, including filter for O'Neill algorithm
