### 2020-06-02 68cf068
Filter kwargs for histogram-based density plots
