### 2020-06-02 2505951
Update changelog, contributors
