### 2020-06-03 #20 @lwasser
hi there @charlesll just checkin in again. are you happy with the review as is as this point? i'd like to ping joss to get the JOSS review started. we are really close to wrapping this up so please respond if you see this ping!
