### 2020-06-04 #20 @charlesll
@lwasser sorry for the delay, yes i'm happy with the new version and the paper.md.
