### 2020-06-05 #20 @morganjwilliams
I'll also add my thanks to @rbeucher and @charlesll for your effort reviewing, and @lwasser for keeping things moving! It's been a trying year for most, and I appreciate the time you've spent on this review (and helping to improve pyrolite in the process!), especially given all that's been happening.

@lwasser yes, that's the version with changes as requested from the reviews. I converted the github tag to a release, and for reference the related Zenodo doi is [10.5281/zenodo.3877690 ](https://doi.org/10.5281/zenodo.3877690). v0.2.8 won't be far away though.

I'll add the badge to the readme with a hotfix patch shortly. Is there any particular branch you'd like to receive pull requests on? I notice there's add-packages and guess this is to be used, but thought it worth checking. Also let me know if for some reason you'd like separate PR's for the package and contributor info.
