### 2020-06-05 94bb28a
Add pyOpenSci to Readme
