### 2020-06-05 #20 @lwasser
thank you @morganjwilliams ! feel free to fork the repo and submit a pr to the master branch. I can test it locally prior to merging. One single PR is just fine! Thank you for asking!!

Congratulations on being accepted by pyOpenSci!! All we need now is the next steps from @arfon  on the JOSS side of things but for pyOpenSci this review is complete. i'll keep the PR open until it runs its way through joss! . And Arfon if you'd like me to direct folks differently (vs pinging you here directly) please say the word!
