### 2020-06-07 #20 @arfon
> This package is going to move on to JOSS for review. i believe @arfon will ask you to implement the items below but let's let him chime in here first!

Yes, this sounds like a good list. Please go ahead and submit this to JOSS and mention that it's a pyOpenSci submission on the form.

> Congratulations on being accepted by pyOpenSci!! All we need now is the next steps from @arfon on the JOSS side of things but for pyOpenSci this review is complete. i'll keep the PR open until it runs its way through joss! . And Arfon if you'd like me to direct folks differently (vs pinging you here directly) please say the word!

This is fine - we can also just have pyOpenSci authors submit without this heads-up directly to me. Submissions from rOpenSci usually just mention this on the form and we act accordingly (authors also usually mention this when the JOSS pre-review issue is opened).
