### 2020-06-08 #20 @lwasser
ok perfect. thank you @arfon i'll add this to our documentation for editors that we can give instructions for folks o submit directly to JOSS and mention that the review happened here already and the package was accepted. Many thanks. So @morganjwilliams please go ahead and submit to JOSS. In your submission, please mention that this went through and was approved by the pyopensci review process. you can even link to this issue. If anything comes up along the way please let me know! i'll close this issue for now but can reopen if need be.

we are actively working on improving our documentation  so it's helpful to know how this part of the process goes IF there are tips that we can provide to maintainers like yourself to make things easier!!
