### 2020-06-09 90dd32f
Modify travis.yml
Modify osx travis config


Modify travis config for osx
