+++ .travis.yml
@@ -1,15 +1,49 @@
 language: python
-os: linux
-dist: bionic
-
 jobs:
   include:
-    - python: 3.6
+    - os: linux
+      dist: bionic
+      python: 3.6
       env: PYVERSION=36
-    - python: 3.7
+    - os: linux
+      dist: bionic
+      python: 3.7
       env: PYVERSION=37
-    - python: 3.8
+    - os: linux
+      dist: bionic
+      python: 3.8
       env: PYVERSION=38
+    - os: osx
+      language: sh
+      env:
+        - HOMEBREW_NO_INSTALL_CLEANUP=1
+        - HOMEBREW_NO_ANALYTICS=1
+        - PYVERSION=36
+      before_cache:
+        # - brew cleanup
+        - rm -f "$HOME/Library/Caches/pip/log/debug.log"
+      cache:
+        directories:
+          # - "$HOME/Library/Caches/Homebrew"
+          - "$HOME/Library/Caches/pip"
+      addons:
+        homebrew:
+          # update: true
+          packages: python3
+      before_install:
+        - python3 -m pip install --upgrade virtualenv
+        - virtualenv -p python3 --system-site-packages "$HOME/venv"
+        - source "$HOME/venv/bin/activate"
+      install:
+        - pip install -q pytest pytest-runner pytest-cov coverage codacy-coverage
+        - pip install coveralls
+        - pip install -e .[skl,stats] # install skl, statsmodels
+      script:
+        - python3 setup.py test
+      after_success:
+        - "ls -lr htmlcov"
+        - coveralls # coverage to coveralls.io
+        - python-codacy-coverage -r coverage.xml
 
 before_install:
   - sudo apt-get install jq curl
