### 2020-06-09 0945c22
Update pyrolite.util.plot for mac renderer bug
