+++ pyrolite/util/plot/__init__.py
@@ -11,10 +11,16 @@ USE_PCOLOR : :class:`bool`
     Option to use the :func:`matplotlib.pyplot.pcolor` function in place
     of :func:`matplotlib.pyplot.pcolormesh`.
 """
+from sys import platform
 import logging
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 logger = logging.getLogger(__name__)
+
+if platform == "darwin":
+    logger.debug('Using TkAgg renderer for Mac.')
+    import matplotlib
+    matplotlib.use('TkAgg')
 
 from .style import DEFAULT_CONT_COLORMAP, DEFAULT_DISC_COLORMAP
 from .density import USE_PCOLOR
