### 2020-06-09 992598f
Allow travis osx failures until 0.2.8
