### 2020-06-09 a060022
Update readme, citation info
