+++ docs/source/conf.py
@@ -313,13 +313,14 @@ rst_prolog = """
 
     {rc}
 
-.. |doibadage| raw:: html
+.. |doibadages| raw:: html
 
     <a style="border-width:0" href="https://doi.org/10.21105/joss.02314">
     <img src="https://joss.theoj.org/papers/10.21105/joss.02314/status.svg" alt="DOI" >
     </a>
+    
     <a href="https://zenodo.org/badge/latestdoi/137172322">
-     <img src="https://zenodo.org/badge/137172322.svg" alt="Archive">
+    <img src="https://zenodo.org/badge/137172322.svg" alt="Archive">
     </a>
 
 """.format(
