### 2020-06-09 1c0859d
Update bibtex citation
