### 2020-06-09 #20 @lwasser
🎆 this was accepted by JOSS. thank you again @arfon !! i'll go ahead and close this issue.
 @rbeucher and @charlesll are you comfortable with my adding you to the pyopensci website as reviewers? it will show your github use profile picture, name the the package you reviewed.

see: https://www.pyopensci.org/contributors/
