### 2020-06-09 #20 @charlesll
Yes, no problem, feel free to do this!

Le mar. 9 juin 2020 à 15:13, Leah Wasser <notifications@github.com> a
écrit :

> 🎆 this was accepted by JOSS. thank you again @arfon
> <https://github.com/arfon> !! i'll go ahead and close this issue.
> @rbeucher <https://github.com/rbeucher> and @charlesll
> <https://github.com/charlesll> are you comfortable with my adding you to
> the pyopensci website as reviewers? it will show your github use profile
> picture, name the the package you reviewed.
>
> see: https://www.pyopensci.org/contributors/
>
