### 2022-09-28 #20 @lwasser
hey there @charlesll @rbeucher  👋 Just a friendly reminder to take 5-10 minutes to fill out our survey . We really appreciate it.  Thank you in advance for helping us by filling out the survey!! 🙌

✨ Morgan thank you so much for taking the time to fill it out 🙌 I will shoot you an email later this week!

[🔗 HERE IS THE SURVEY LINK 🔗 ](https://forms.gle/ZZXcD1hr18tYmY2p6)
